requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-aiplatform
vertexai
pydantic
httpx
